# docs/swagger.yaml
definitions:
  models.Project:
    properties:
      id:
        example: 1
        type: integer
      name:
        example: Work
        type: string
      user_id:
        example: 2
        type: integer
    type: object
  models.RegisterRequest:
    properties:
      password:
        example: password123
        type: string
      role:
        description: optional, defaults to "user"
        example: admin
        type: string
      username:
        example: sumit
        type: string
    type: object
  models.Tag:
    properties:
      id:
        example: 1
        type: integer
      name:
        example: urgent
        type: string
    type: object
  models.Task:
    properties:
      created_at:
        example: "2025-05-07T12:34:56Z"
        type: string
      id:
        example: 1
        type: integer
      project_id:
        example: 1
        type: integer
      status:
        example: in-progress
        type: string
      tags:
        items:
          $ref: '#/definitions/models.Tag'
        type: array
      title:
        example: Buy milk
        type: string
      updated_at:
        example: "2025-05-07T13:34:56Z"
        type: string
      user_id:
        example: 2
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a task manager backend built with Go.
  title: Task Manager API
  version: "1.0"
paths:
  /projects:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Project'
            type: array
      security:
      - BearerAuth: []
      summary: Get all projects
      tags:
      - Projects
    post:
      consumes:
      - application/json
      parameters:
      - description: Project data
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/models.Project'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Project'
      security:
      - BearerAuth: []
      summary: Create a new project
      tags:
      - Projects
  /projects/{id}/tasks:
    get:
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Task'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get tasks for a specific project
      tags:
      - Projects
  /register:
    post:
      consumes:
      - application/json
      parameters:
      - description: User credentials with optional role (admin or user)
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user (optionally as admin)
      tags:
      - Auth
  /tags:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Tag'
            type: array
      summary: Get all tags
      tags:
      - Tags
    post:
      consumes:
      - application/json
      parameters:
      - description: Tag info
        in: body
        name: tag
        required: true
        schema:
          $ref: '#/definitions/models.Tag'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Tag'
      summary: Create a new tag
      tags:
      - Tags
  /tasks:
    get:
      parameters:
      - description: Filter by Project ID
        in: query
        name: project_id
        type: integer
      - description: Max number of results
        in: query
        name: limit
        type: integer
      - description: Number of results to skip
        in: query
        name: offset
        type: integer
      - description: Sort by field (e.g. created_at, title, status)
        in: query
        name: sort
        type: string
      - description: Sort order (asc or desc)
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Task'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all tasks for the logged-in user (with filters, pagination, sorting)
      tags:
      - Tasks
    post:
      consumes:
      - application/json
      parameters:
      - description: Task info
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/models.Task'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Task'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new task
      tags:
      - Tasks
  /tasks/{id}:
    delete:
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a task
      tags:
      - Tasks
    put:
      consumes:
      - application/json
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated task
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/models.Task'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Task'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a task
      tags:
      - Tasks
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"